--- main.py
@@ -1,5 +1,0 @@
-def main():
-    print("Hello World!")
-
-if __name__ == "__main__":
-    main()
